+++ CHANGES.md
@@ -1,18 +1,22 @@
 Change history of go-restful
 =
 
+2013-05-22
+	
+ - (api add) Added support for request/response filter functions
+
 2013-05-18
--
+
 
  - (api add) Added feature to change the default Http Request Dispatch function (travis cline)
  - (api change) Moved Swagger Webservice to swagger package (see example restful-user)
 
 [2012-11-14 .. 2013-05-18>
-- 
+ 
  - See https://github.com/emicklei/go-restful/commits
 
 2012-11-14
--
+
  - Initial commit